Update ViewConfiguration.java
The maximum fling velocity is a bit too much at 16000. There's much more stuff that has to be loaded on the screen with a single fling and that causes serious stutters, probably because the CPU can't load very well so much stuff in such a short amount time. 10000 is a bit increased over the stock value but it should look a bit less stuttery than double.

Signed-off-by: spezi77 <spezi77@gmx.com>